[ticket/12291] Improve and move the migration file. Fix smilies/icons paths.
PHPBB3-12291
